\name{eQTLcross-class}
\docType{class}
\alias{eQTLcross-class}
\alias{eQTLcrossParam-class}
\alias{eQTLcross}
\alias{eQTLcrossParam}
\alias{eQTLcross,map,matrix,HMgmm-method}
\alias{eQTLcross,map,matrix,graphBAM-method}
\alias{eQTLcross,map,matrix,matrix-method}
\alias{eQTLcross,map,missing,HMgmm-method}
\alias{eQTLcross,map,missing,matrix-method}
\alias{eQTLcross,map,missing,missing-method}
\alias{addGenes,eQTLcross,missing-method}
\alias{addGenes,eQTLcross,numeric-method}
\alias{addGenes,eQTLcross,integer-method}
\alias{addGenes}
\alias{addeQTL,eQTLcross-method}
\alias{addeQTL}
\alias{addGeneAssociation,eQTLcross-method}
\alias{addGeneAssociation}
\alias{reQTLcross,eQTLcrossParam,missing-method}
\alias{reQTLcross,missing,eQTLcrossParam-method}
\alias{reQTLcross,numeric,eQTLcrossParam-method}
\alias{reQTLcross,integer,eQTLcrossParam-method}
\alias{reQTLcross,eQTLcross,missing-method}
\alias{reQTLcross,missing,eQTLcross-method}
\alias{reQTLcross,numeric,eQTLcross-method}
\alias{reQTLcross,integer,eQTLcross-method}
\alias{reQTLcross}
\alias{names,eQTLcross-method}
\alias{$,eQTLcross-method}
\alias{show,eQTLcross-method}
\alias{show,eQTLcrossParam-method}
\alias{plot,eQTLcross,ANY-method}
\alias{alleQTL,eQTLcross-method}
\alias{alleQTL}
\alias{ciseQTL,eQTLcross,missing-method}
\alias{ciseQTL,eQTLcross,numeric-method}
\alias{ciseQTL}
\alias{transeQTL,eQTLcross,missing-method}
\alias{transeQTL,eQTLcross,numeric-method}
\alias{transeQTL}
\alias{sim.cross,map,eQTLcross-method}
\alias{sim.cross,map,matrix-method}
\alias{sim.cross}

\title{eQTL experimental cross model class}
\description{
The expression quantitative trait loci (eQTL) experimental cross model class serves
the purpose of holding all necessary information to simulate genetical genomics data
from an experimental cross.
}
\author{R. Castelo}
\keyword{classes}
