\name{eQTLnetwork-class}
\docType{class}
\alias{eQTLnetwork-class}
\alias{eQTLnetwork}
\alias{show,eQTLnetwork-method}
\alias{plot,eQTLnetwork,ANY-method}
\alias{alleQTL,eQTLnetwork-method}
\alias{alleQTL}
\alias{ciseQTL,eQTLnetwork,missing-method}
\alias{ciseQTL,eQTLnetwork,numeric-method}
\alias{ciseQTL}
\alias{allGeneAssociations,eQTLnetwork-method}
\alias{allGeneAssociations}
\alias{resetCutoffs,eQTLnetwork-method}
\alias{resetCutoffs}
\alias{varExplained,eQTLnetworkEstimationParam,eQTLnetwork-method}
\alias{varExplained}
\alias{geneNames,eQTLnetwork-method}
\alias{markerNames,eQTLnetwork-method}
\alias{geneticMap,eQTLnetwork-method}
\alias{physicalMap,eQTLnetwork-method}
\alias{geneAnnotation,eQTLnetwork-method}
\alias{graph,eQTLnetwork-method}
\alias{graph}


\title{eQTL network model class}
\description{
The expression quantitative trait loci (eQTL) network class serves
the purpose of holding the result of estimating an eQTL network from
genetical genomics data.
}
\author{R. Castelo}
\seealso{
  \code{\link{eQTLnetworkEstimationParam-class}}
  \code{\link{eQTLcross-class}}
}
\keyword{classes}
