\name{eQTLnetworkEstimate}
\alias{eQTLnetworkEstimate}
\alias{eQTLnetworkEstimate,eQTLnetworkEstimationParam,formula,missing-method}
\alias{eQTLnetworkEstimate,eQTLnetworkEstimationParam,formula,eQTLnetwork-method}
\alias{eQTLnetworkEstimate,eQTLnetworkEstimationParam,missing,eQTLnetwork-method}

\title{Estimation of a eQTL network from genetical genomics data}
\description{
The methods \code{eQTLnetworkEstimate()} perform the estimation of eQTL
networks from genetical genomics data using as input an
\code{\link{eQTLnetworkEstimationParam}} object and, eventually, a
\code{\link{eQTLnetwork-class}} object.
}
\author{R. Castelo}
\seealso{
  \code{\link{eQTLnetworkEstimationParam}}
  \code{\link{eQTLnetwork-class}}
}
\keyword{classes}
