\name{eQTLnetworkEstimationParam-class}
\docType{class}
\alias{eQTLnetworkEstimationParam-class}
\alias{eQTLnetworkEstimationParam}
\alias{show,eQTLnetworkEstimationParam-method}
\alias{geneticMap,eQTLnetworkEstimationParam-method}
\alias{geneticMap}
\alias{physicalMap,eQTLnetworkEstimationParam-method}
\alias{physicalMap}
\alias{ggData,eQTLnetworkEstimationParam-method}
\alias{ggData}
\alias{markerNames,eQTLnetworkEstimationParam-method}
\alias{markerNames}
\alias{geneNames,eQTLnetworkEstimationParam-method}
\alias{geneNames}
\alias{geneAnnotation,eQTLnetworkEstimationParam-method}
\alias{geneAnnotation}


\title{eQTL network parameter model class}
\description{
The expression quantitative trait loci (eQTL) network parameter class serves
the purpose of holding the parameter information to estimate
an eQTL network from genetical genomics data with the function
\code{\link{eQTLnetworkEstimate}}.
}
\author{R. Castelo}
\seealso{
  \code{\link{eQTLnetworkEstimate}}
  \code{\link{eQTLnetwork-class}}
}
\keyword{classes}
