\name{graphParam-class}
\docType{class}
\alias{graphParam-class}
\alias{markedGraphParam-class}
\alias{erGraphParam-class}
\alias{dRegularGraphParam-class}
\alias{erMarkedGraphParam-class}
\alias{dRegularMarkedGraphParam-class}
\alias{erGraphParam}
\alias{dRegularGraphParam}
\alias{erMarkedGraphParam}
\alias{dRegularMarkedGraphParam}
\alias{rgraphBAM,graphParam,missing-method}
\alias{rgraphBAM,missing,graphParam-method}
\alias{rgraphBAM,numeric,graphParam-method}
\alias{rgraphBAM,integer,graphParam-method}
\alias{rgraphBAM,erGraphParam,missing-method}
\alias{rgraphBAM,missing,erGraphParam-method}
\alias{rgraphBAM,numeric,erGraphParam-method}
\alias{rgraphBAM,integer,erGraphParam-method}
\alias{rgraphBAM,dRegularGraphParam,missing-method}
\alias{rgraphBAM,missing,dRegularGraphParam-method}
\alias{rgraphBAM,numeric,dRegularGraphParam-method}
\alias{rgraphBAM,integer,dRegularGraphParam-method}
\alias{rgraphBAM}
\alias{show,graphParam-method}
\alias{plot,graphBAM,ANY-method}

\title{Graph parameter classes}
\description{
Graph parameter classes are defined to ease the simulation of different
types of graphs by using a single interface \code{rgraphBAM()}.
}
\author{R. Castelo}
\keyword{classes}
