\name{eQTLnetworkEstimationParam-class}
\docType{class}
\alias{eQTLnetworkEstimationParam-class}
\alias{eQTLnetworkEstimationParam}
\alias{show,eQTLnetworkEstimationParam-method}

\title{eQTL network estimation parameter classes}
\description{
The eQTL network estimation parameter class serves the purpose of
facilitating the call to the corresponding functions in the \code{qpgraph}
package to estimate an eQTL network from genetical genomics data. This is
still currently experimental und under development.
}
\author{R. Castelo}
\keyword{classes}
