\name{qpSampleFromHMGM}
\alias{qpSampleFromHMGM}

\title{
Sample from homogeneous mixed graphical Markov models
}
\description{
Samples synthetic data from homogeneous mixed graphical Markov models (experimental feature).
}
\usage{
qpSampleFromHMGM <- function(n=10, hmgm=qpRndHMGM())
}
\arguments{
  \item{n}{number of observations to sample.}
  \item{hmgm}{homogeneous mixed graphical Markov model as generated by the function \code{\link{qpRndHMGM}}.}
}
\details{
This function samples synthetic data from a random homogeneous mixed graphical model
build with the function \code{\link{qpRndHMGM}}.
This is still an experimental feature.

}
\value{
The sampled synthetic data.
}
\references{
Castelo, R. and Roverato, A. A robust procedure for
Gaussian graphical model search from microarray data with p larger than n,
\emph{J. Mach. Learn. Res.}, 7:2621-2650, 2006.
}
\author{R. Castelo and A. Roverato}
\seealso{
  \code{\link{qpRndGraph}}
  \code{\link{qpSampleFromHMGM}}
}
\examples{
qpSampleFromHMGM()
}
\keyword{models}
\keyword{multivariate}
